eshte bere login customer i cili mund te i shoh  vetem produktet e veta qe i bon ne cart,admini i sheh te gjitha order,slider,product

+++ app/Http/Controllers/OrdersController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Order;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class OrdersController extends Controller
 {
@@ -14,7 +15,12 @@ class OrdersController extends Controller
      */
     public function index()
     {
-        $orders = Order::all();
+        if (!Auth::user()->hasRole('admin')) {  // e bojm nese nuk osht auth si admin mos ti sheh orderat perndryshe le ti sheh vetem orderat e vete
+            $orders = Order::where('user_id', Auth::id())->get();
+        } else {
+            $orders = Order::all();
+        }
+
         return view('dashboard.orders.index',compact('orders'));
     }
 
@@ -83,6 +89,10 @@ class OrdersController extends Controller
      */
     public function destroy($id)
     {
+        if (!Auth::user()->hasRole('admin')) {   //e kemi bo nese nuk osht admin mos me mujt mi fshi orderat
+            # code...
+            return redirect()->route('orders.index')->with('status','You are not allowed to delete orders!');
+        }
         $order = Order::findOrFail($id);
         $order->delete();
